+++ bcipy/signal/tests/model/pca_rda_kde/test_pca_rda_kde.py
@@ -9,9 +9,9 @@ import pytest
 from scipy.stats import norm
 
 from bcipy.exceptions import SignalException
-from bcipy.helpers.load import load_signal_models
-from bcipy.helpers.save import save_model
-from bcipy.helpers.symbols import alphabet
+from bcipy.io.load import load_signal_models
+from bcipy.io.save import save_model
+from bcipy.core.symbols import alphabet
 from bcipy.signal.model import ModelEvaluationReport, PcaRdaKdeModel
 from bcipy.signal.model.classifier import RegularizedDiscriminantAnalysis
 from bcipy.signal.model.cross_validation import cross_validation