+++ bcipy/task/control/evidence.py
@@ -8,7 +8,7 @@ import numpy as np
 from bcipy.acquisition.multimodal import ContentType
 from bcipy.config import SESSION_LOG_FILENAME
 from bcipy.helpers.acquisition import analysis_channels
-from bcipy.helpers.stimuli import TrialReshaper
+from bcipy.core.stimuli import TrialReshaper
 from bcipy.signal.model import SignalModel
 from bcipy.task.data import EvidenceType
 from bcipy.task.exceptions import MissingEvidenceEvaluator